chore: merge vm_operator and vm_cluster into victoria_metrics
Summary:
Users won't want to know or care. This should help the configuration
make more sense too

Test Plan:
`bix ex-test-deep`
