--- platform_umbrella/apps/common_core/lib/common_core/batteries/vm_operator_config.ex
@@ -1,11 +1,0 @@
-defmodule CommonCore.Batteries.VMOperatorConfig do
-  @moduledoc false
-
-  use CommonCore, :embedded_schema
-
-  alias CommonCore.Defaults
-
-  batt_polymorphic_schema type: :vm_operator do
-    defaultable_field :vm_operator_image, :string, default: Defaults.Images.vm_operator_image()
-  end
-end
